fixme: Add back NativeModuleLoader CompileAndCall
Node removed the NativeModuleLoader CompileAndCall function and exposed
the LookupAndCompile function, with the intention of giving the ability
to the caller to handle errors. The node PR made the changes
https://github.com/nodejs/node/pull/25667.
I'm adding this back for now, but this can easily be moved to electron
as a helper or can probably be upstreamed somehow.